chore: extend tests to cover the RedisOauthTokenStore (#2)
