+++ src/test/java/org/apache/sling/auth/oauth_client/impl/JcrUserHomeOAuthTokenStoreTest.java
@@ -18,199 +18,37 @@ package org.apache.sling.auth.oauth_client.impl;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import java.util.concurrent.TimeUnit;
-
 import javax.jcr.RepositoryException;
 
 import org.apache.jackrabbit.api.security.user.User;
 import org.apache.sling.api.resource.Resource;
 import org.apache.sling.api.resource.ValueMap;
 import org.apache.sling.auth.oauth_client.ClientConnection;
-import org.apache.sling.auth.oauth_client.OAuthToken;
-import org.apache.sling.auth.oauth_client.TokenState;
-import org.apache.sling.auth.oauth_client.impl.Converter;
-import org.apache.sling.auth.oauth_client.impl.JcrUserHomeOAuthTokenStore;
-import org.apache.sling.jackrabbit.usermanager.impl.AuthorizableAdapterFactory;
 import org.apache.sling.testing.mock.sling.ResourceResolverType;
 import org.apache.sling.testing.mock.sling.junit5.SlingContext;
-import org.apache.sling.testing.mock.sling.junit5.SlingContextExtension;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
 
-import com.nimbusds.jwt.JWTClaimsSet;
-import com.nimbusds.jwt.PlainJWT;
-import com.nimbusds.oauth2.sdk.token.BearerAccessToken;
-import com.nimbusds.oauth2.sdk.token.RefreshToken;
 import com.nimbusds.openid.connect.sdk.token.OIDCTokens;
 
-@ExtendWith(SlingContextExtension.class)
-class JcrUserHomeOAuthTokenStoreTest {
-
-    private final SlingContext context = new SlingContext(ResourceResolverType.JCR_OAK);
-    private final MockOidcConnection connection = MockOidcConnection.DEFAULT_CONNECTION;
-
-    @BeforeEach
-    public void registerAdapterFactories() {
-        context.registerInjectActivateService(new AuthorizableAdapterFactory());
+class JcrUserHomeOAuthTokenStoreTest extends TokenStoreTestSupport<JcrUserHomeOAuthTokenStore> {
+    
+    JcrUserHomeOAuthTokenStoreTest() {
+        super(MockOidcConnection.DEFAULT_CONNECTION, new SlingContext(ResourceResolverType.JCR_OAK));
     }
 
-    @Test
-    void persistTokens_accessTokenOnly() throws RepositoryException {
-
-        OIDCTokens tokens = new OIDCTokens(new BearerAccessToken(12), null);
-
-        JcrUserHomeOAuthTokenStore tokenStore = new JcrUserHomeOAuthTokenStore();
-        tokenStore.persistTokens(connection, context.resourceResolver(), Converter.toSlingOAuthTokens(tokens));
-
+    @Override
+    JcrUserHomeOAuthTokenStore createTokenStore() {
+        return new JcrUserHomeOAuthTokenStore();
+    }
+    
+    @Override
+    protected void getAccessToken_valid_postCheck(OIDCTokens input) throws RepositoryException {
         Resource connectionResource = getConnectionResource(connection);
 
         ValueMap connectionProps = connectionResource.getValueMap();
         assertThat(connectionProps)
             .as("stored tokens for connection")
             .containsOnlyKeys("jcr:primaryType", "access_token")
-            .containsEntry("access_token", tokens.getAccessToken().getValue());
-    }
-
-    @Test
-    void persistTokens_accessAndIdToken() throws RepositoryException {
-
-        OIDCTokens tokens = new OIDCTokens(new PlainJWT(new JWTClaimsSet.Builder().issuer("example.com").build()), new BearerAccessToken(12), null);
-
-        JcrUserHomeOAuthTokenStore tokenStore = new JcrUserHomeOAuthTokenStore();
-        tokenStore.persistTokens(connection, context.resourceResolver(), Converter.toSlingOAuthTokens(tokens));
-
-        Resource connectionResource = getConnectionResource(connection);
-
-        // we explicitly do not store the id token, that is only useful if we want an authentication handler
-        ValueMap connectionProps = connectionResource.getValueMap();
-        assertThat(connectionProps)
-            .as("stored tokens for connection")
-            .containsOnlyKeys("jcr:primaryType", "access_token")
-            .containsEntry("access_token", tokens.getAccessToken().getValue());
-    }
-
-    
-    @Test
-    void getAccessToken_missing() {
-        
-        JcrUserHomeOAuthTokenStore tokenStore = new JcrUserHomeOAuthTokenStore();
-        
-        OAuthToken accessToken = tokenStore.getAccessToken(connection, context.resourceResolver());
-        
-        assertThat(accessToken).as("access token")
-            .isNotNull()
-            .extracting( OAuthToken::getState )
-            .isEqualTo( TokenState.MISSING );
-    }
-    
-    @Test
-    void getAccessToken_valid() {
-        
-        OIDCTokens tokens = new OIDCTokens(new BearerAccessToken(12), null);
-
-        JcrUserHomeOAuthTokenStore tokenStore = new JcrUserHomeOAuthTokenStore();
-        tokenStore.persistTokens(connection, context.resourceResolver(), Converter.toSlingOAuthTokens(tokens));
-        
-        OAuthToken accessToken = tokenStore.getAccessToken(connection, context.resourceResolver());
-        assertThat(accessToken).as("access token")
-            .isNotNull()
-            .extracting( OAuthToken::getState , OAuthToken::getValue )
-            .containsExactly( TokenState.VALID, tokens.getAccessToken().getValue() );
-    }
-    
-    @Test
-    void getAccessToken_notYetExpired() throws InterruptedException {
-        
-        OIDCTokens tokens = new OIDCTokens(new BearerAccessToken(12, 3600, null), null);
-        
-        JcrUserHomeOAuthTokenStore tokenStore = new JcrUserHomeOAuthTokenStore();
-        tokenStore.persistTokens(connection, context.resourceResolver(), Converter.toSlingOAuthTokens(tokens));
-
-        OAuthToken accessToken = tokenStore.getAccessToken(connection, context.resourceResolver());
-        assertThat(accessToken).as("access token")
-            .isNotNull()
-            .extracting( OAuthToken::getState )
-            .isEqualTo( TokenState.VALID);
-    }    
-
-    @Test
-    void getAccessToken_expired() throws InterruptedException {
-        
-        int lifetimeSeconds = 1;
-        OIDCTokens tokens = new OIDCTokens(new BearerAccessToken(12, lifetimeSeconds, null), null);
-        
-        JcrUserHomeOAuthTokenStore tokenStore = new JcrUserHomeOAuthTokenStore();
-        tokenStore.persistTokens(connection, context.resourceResolver(), Converter.toSlingOAuthTokens(tokens));
-
-        // wait for the token to expire
-        Thread.sleep( TimeUnit.SECONDS.toMillis( 2 * lifetimeSeconds ) );
-        
-        OAuthToken accessToken = tokenStore.getAccessToken(connection, context.resourceResolver());
-        assertThat(accessToken).as("access token")
-            .isNotNull()
-            .extracting( OAuthToken::getState )
-            .isEqualTo( TokenState.EXPIRED );
-    }
-    
-    @Test
-    void getRefreshToken_valid() {
-        OIDCTokens tokens = new OIDCTokens(new BearerAccessToken(12), new RefreshToken(12));
-
-        JcrUserHomeOAuthTokenStore tokenStore = new JcrUserHomeOAuthTokenStore();
-        tokenStore.persistTokens(connection, context.resourceResolver(), Converter.toSlingOAuthTokens(tokens));
-        
-        OAuthToken refreshToken = tokenStore.getRefreshToken(connection, context.resourceResolver());
-        assertThat(refreshToken).as("refresh token")
-            .isNotNull()
-            .extracting( OAuthToken::getState , OAuthToken::getValue )
-            .containsExactly( TokenState.VALID, tokens.getRefreshToken().getValue() );
-    }
-    
-    @Test
-    void getRefreshToken_missing() {
-        
-        JcrUserHomeOAuthTokenStore tokenStore = new JcrUserHomeOAuthTokenStore();
-        
-        OAuthToken refreshToken = tokenStore.getRefreshToken(connection, context.resourceResolver());
-        assertThat(refreshToken).as("refresh token")
-            .isNotNull()
-            .extracting( OAuthToken::getState )
-            .isEqualTo( TokenState.MISSING);
-    }
-    
-    @Test
-    void clearAccessToken() {
-        JcrUserHomeOAuthTokenStore tokenStore = new JcrUserHomeOAuthTokenStore();
-            
-        OIDCTokens tokens = new OIDCTokens(new BearerAccessToken(12), null);
-        tokenStore.persistTokens(connection, context.resourceResolver(), Converter.toSlingOAuthTokens(tokens));
-
-        assertThat(tokenStore.getAccessToken(connection, context.resourceResolver()))
-            .as("persisted access token")
-            .isNotNull()
-            .extracting( OAuthToken::getState )
-            .isEqualTo( TokenState.VALID );
-
-        tokenStore.clearAccessToken(connection, context.resourceResolver());
-        
-        assertThat(tokenStore.getAccessToken(connection, context.resourceResolver()))
-            .as("cleared access token")
-            .isNotNull()
-            .extracting( OAuthToken::getState )
-            .isEqualTo( TokenState.MISSING );
-    }
-    
-    @Test
-    void clearAccessToken_missingAlready() {
-        JcrUserHomeOAuthTokenStore tokenStore = new JcrUserHomeOAuthTokenStore();
-        tokenStore.clearAccessToken(connection, context.resourceResolver());
-        
-        assertThat(tokenStore.getAccessToken(connection, context.resourceResolver()))
-            .as("access token")
-            .isNotNull()
-            .extracting( OAuthToken::getState )
-            .isEqualTo( TokenState.MISSING );
+            .containsEntry("access_token", input.getAccessToken().getValue());        
     }
     
     private Resource getConnectionResource(ClientConnection connection) throws RepositoryException {
